tests/requirements.txt
# Notification Service - Test Framework Requirements
# Python 3.11+

# Core Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-html==4.1.1
pytest-xdist==3.5.0
pytest-timeout==2.2.0
pytest-ordering==0.6
pytest-repeat==0.9.3

# HTTP Client
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Data Validation & Generation
pydantic==2.5.2
faker==21.0.0
factory-boy==3.3.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Assertions
assertpy==1.1
deepdiff==6.7.1

# Reporting
allure-pytest==2.13.2
rich==13.7.0

# Performance Testing
locust==2.20.1

# Security Testing
python-jose==3.3.0

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
jsonschema==4.20.0
pyyaml==6.0.1

# Code Quality
black==23.12.0
isort==5.13.2
flake8==6.1.0
mypy==1.7.1
